+++ tests/test_token.py
@@ -11,6 +11,7 @@ from oic.utils.keyio import KeyJar
 from oic.utils.sdb import AccessCodeUsed
 from oic.utils.sdb import AuthnEvent
 from oic.utils.sdb import DefaultToken
+from oic.utils.sdb import DictSessionBackend
 from oic.utils.sdb import ExpiredToken
 from oic.utils.sdb import SessionDB
 
@@ -190,7 +191,7 @@ class TestSessionDB(object):
 
         self.sdb = SessionDB(
             "https://example.com/",
-            db={},
+            db=DictSessionBackend(),
             code_factory=DefaultToken(
                 'supersecret', 'verybadpassword', typ='A', lifetime=600),
             token_factory=JWTToken('T', keyjar=kj,